+++ src/main/java/org/jabref/gui/search/GlobalSearchBar.java
@@ -5,12 +5,9 @@ import java.util.List;
 import java.util.Optional;
 import java.util.regex.Pattern;
 import java.util.regex.PatternSyntaxException;
-import java.util.stream.Collectors;
-
 import javafx.application.Platform;
 import javafx.beans.binding.Bindings;
 import javafx.beans.binding.BooleanBinding;
-import javafx.collections.ObservableList;
 import javafx.css.PseudoClass;
 import javafx.event.Event;
 import javafx.geometry.Insets;
@@ -41,12 +38,9 @@ import org.jabref.gui.Globals;
 import org.jabref.gui.JabRefFrame;
 import org.jabref.gui.StateManager;
 import org.jabref.gui.autocompleter.AppendPersonNamesStrategy;
-import org.jabref.gui.autocompleter.AutoCompleteFirstNameMode;
 import org.jabref.gui.autocompleter.AutoCompletionTextInputBinding;
-import org.jabref.gui.autocompleter.PersonNameStringConverter;
 import org.jabref.gui.autocompleter.SuggestionProvider;
 import org.jabref.gui.externalfiletype.ExternalFileTypes;
-import org.jabref.gui.groups.GroupViewMode;
 import org.jabref.gui.icon.IconTheme;
 import org.jabref.gui.keyboard.KeyBinding;
 import org.jabref.gui.keyboard.KeyBindingRepository;
@@ -58,14 +52,7 @@ import org.jabref.gui.util.IconValidationDecorator;
 import org.jabref.gui.util.TooltipTextUtil;
 import org.jabref.logic.l10n.Localization;
 import org.jabref.logic.search.SearchQuery;
-import org.jabref.logic.util.io.FileUtil;
-import org.jabref.model.database.BibDatabaseContext;
 import org.jabref.model.entry.Author;
-import org.jabref.model.entry.BibEntry;
-import org.jabref.model.entry.field.UnknownField;
-import org.jabref.model.groups.GroupTreeNode;
-import org.jabref.model.search.matchers.MatcherSet;
-import org.jabref.model.search.matchers.MatcherSets;
 import org.jabref.preferences.PreferencesService;
 import org.jabref.preferences.SearchPreferences;
 
@@ -107,7 +94,6 @@ public class GlobalSearchBar extends HBox {
 
     private SearchPreferences searchPreferences;
 
-    private final GroupViewMode groupViewMode;
     private final GlobalSearchResultDialog globalSearchDialog;
 
     public GlobalSearchBar(JabRefFrame frame, StateManager stateManager, PreferencesService preferencesService, CountingUndoManager undoManager, DialogService dialogService) {
@@ -115,8 +101,6 @@ public class GlobalSearchBar extends HBox {
         this.stateManager = stateManager;
         this.preferencesService = preferencesService;
         this.searchPreferences = preferencesService.getSearchPreferences();
-
-        groupViewMode = preferencesService.getGroupViewMode();
 
         searchField.disableProperty().bind(needsDatabase(stateManager).not());
 
@@ -199,8 +183,7 @@ public class GlobalSearchBar extends HBox {
             searchQuery.ifPresent(query -> {
                 updateResults(this.stateManager.getSearchResultSize().intValue(), SearchDescribers.getSearchDescriberFor(query).getDescription(),
                               query.isGrammarBasedSearch());
-                updateGlobalSearchResults();
-
+                globalSearchDialog.doGlobalSearch();
             });
         });
 
@@ -295,53 +278,7 @@ public class GlobalSearchBar extends HBox {
 
     }
 
-    private void updateGlobalSearchResults() {
-        if (stateManager.isGlobalSearchActive()) {
-            BibDatabaseContext context = new BibDatabaseContext();
 
-            for (var db : this.stateManager.getOpenDatabases()) {
-
-                List<BibEntry> result = db.getEntries().stream().filter(entry -> isMatched(stateManager.activeGroupProperty(), stateManager.activeSearchQueryProperty().get(), entry))
-                                          .map(s -> {
-                                              BibEntry newEntry = s.withField(new UnknownField(GlobalSearchResultDialog.LIBRARY_NAME_FIELD), FileUtil.getBaseName(db.getDatabasePath().orElse(null)));
-                                              return newEntry;
-                                          })
-                                          .collect(Collectors.toList());
-
-                context.getDatabase().insertEntries(result);
-            }
-          this.globalSearchDialog.addEntriesToBibContext(context);
-        }
-    }
-
-    private boolean isMatched(ObservableList<GroupTreeNode> groups, Optional<SearchQuery> query, BibEntry entry) {
-        return isMatchedByGroup(groups, entry) && isMatchedBySearch(query, entry);
-    }
-
-    private boolean isMatchedBySearch(Optional<SearchQuery> query, BibEntry entry) {
-        return query.map(matcher -> matcher.isMatch(entry))
-                    .orElse(true);
-    }
-
-    private boolean isMatchedByGroup(ObservableList<GroupTreeNode> groups, BibEntry entry) {
-        return createGroupMatcher(groups)
-                                         .map(matcher -> matcher.isMatch(entry))
-                                         .orElse(true);
-    }
-
-    private Optional<MatcherSet> createGroupMatcher(List<GroupTreeNode> selectedGroups) {
-        if ((selectedGroups == null) || selectedGroups.isEmpty()) {
-            // No selected group, show all entries
-            return Optional.empty();
-        }
-
-        final MatcherSet searchRules = MatcherSets.build(groupViewMode == GroupViewMode.INTERSECTION ? MatcherSets.MatcherType.AND : MatcherSets.MatcherType.OR);
-
-        for (GroupTreeNode node : selectedGroups) {
-            searchRules.addRule(node.getSearchMatcher());
-        }
-        return Optional.of(searchRules);
-    }
 
     private boolean validRegex() {
         try {
@@ -365,9 +302,7 @@ public class GlobalSearchBar extends HBox {
     public void setAutoCompleter(SuggestionProvider<Author> searchCompleter) {
         if (preferencesService.getAutoCompletePreferences().shouldAutoComplete()) {
             AutoCompletionTextInputBinding<Author> autoComplete = AutoCompletionTextInputBinding.autoComplete(searchField,
-                                                                                                              searchCompleter::provideSuggestions,
-                                                                                                              new PersonNameStringConverter(false, false, AutoCompleteFirstNameMode.BOTH),
-                                                                                                              new AppendPersonNamesStrategy());
+                                                                                                              searchCompleter::provideSuggestions,                                                                                       new AppendPersonNamesStrategy());
             AutoCompletePopup<Author> popup = getPopup(autoComplete);
             popup.setSkin(new SearchPopupSkin<>(popup));
         }
